Estimating Rotemberg Weights for a Bartik "Instrument"

bw() estimates the Rotemberg weights for a Bartik "instrument" outlined in Goldsmith-Pinkham, Sorkin and Swift (2018) and the just-identified IV estimates using local industry shares, which are the actual instruments.

bw(master, y, x, controls = NULL, weight = NULL, local, Z, global, G)

Arguments

master
The master data frame.

y
A string for outcome variable. It should be a variable in master.

x
A string for the endogenous variable. It should be a variable in master.

controls
A string or character vector for the control variables. controls are optional and should be in master.

weight
A string for the weighted variable. weight is optional and should be in master.

local
The local data frame. It should be in wide format.

Z
A string or character vector for the the local industry shares.

global
The global data frame.

G
A string for the the overall industry growth rates.

Value

A data frame built from global with two additional variables: alpha (the Rotemberg weights), and beta (the just-identified IV estimates).

Details

The typical Bartik "instrument" has two components: the local industry shares (usually at the level of location-industry or location-year-industry), and the overall industry growth rates (usually at the level of industry or industry-year). The outcome variable and the endogenous variable are at the level of location or location-year.

Because the key variables are at three different levels, bw() proceeds by taking three different datasets: (1) a main master data frame containing the dependent variable (y), the causal variable of interest (x), a set of control variables (controls), and the weighted variables (weight); (2) a local data frame containing the set of local industry shares (Z); and (3) a global data frame containing the overall industry growth rates (G). At the moment, it is necessary to transform the "local" dataset from long format to wide format.

Examples

library(bartik.weight)

index = c("czone", "year")
y = "d_sh_empl_mfg"
x = "d_tradeusch_pw"
controls = c("reg_midatl", "reg_encen", "reg_wncen", "reg_satl",
  "reg_escen", "reg_wscen", "reg_mount", "reg_pacif", "l_sh_popedu_c",
  "l_sh_popfborn", "l_sh_empl_f", "l_sh_routine33", "l_task_outsource",
  "t2", "l_shind_manuf_cbp")
weight = "timepwt48"
Z = setdiff(names(ADH_local_wide), index)
G = "trade_"

bw(ADH_master, y, x, controls, weight, ADH_local_wide, Z, ADH_global, G)

#> # A tibble: 780 x 5
#>     year ind     trade_        alpha     beta
#>    <dbl> <chr>    <dbl>        <dbl>    <dbl>
#>  1 1990. 2011   1.21    -0.00102      -2.80  
#>  2 1990. 2015   7.46    -0.0104       -1.06  
#>  3 1990. 2022   0.00698  0.000000663   7.74  
#>  4 1990. 2023   3.16     0.0000676     8.82  
#>  5 1990. 2024   0.       0.            3.93  
#>  6 1990. 2026   0.0333   0.0000218    -0.0432
#>  7 1990. 2032   0.176    0.00000474   13.6   
#>  8 1990. 2033   3.78     0.0000286    29.0   
#>  9 1990. 2034  12.2     -0.00188       0.923 
#> 10 1990. 2035   5.11    -0.000357      1.59  
#> # ... with 770 more rows